+++ core/src/androidMain/kotlin/com/revenuecat/purchases/kmp/PurchasesDelegate.android.kt
@@ -1,5 +1,7 @@
 package com.revenuecat.purchases.kmp
 
+import com.revenuecat.purchases.interfaces.UpdatedCustomerInfoListener
+import com.revenuecat.purchases.kmp.mappings.toCustomerInfo
 import com.revenuecat.purchases.kmp.models.CustomerInfo
 import com.revenuecat.purchases.kmp.models.PurchasesError
 import com.revenuecat.purchases.kmp.models.StoreProduct
@@ -21,5 +23,18 @@ public class UpdatedCustomerInfoDelegate(
         startPurchase: (onError: (error: PurchasesError, userCancelled: Boolean) -> Unit, onSuccess: (storeTransaction: StoreTransaction, customerInfo: CustomerInfo) -> Unit) -> Unit
     ) {
         // No-op on Android.
+    }
+}
+
+internal fun UpdatedCustomerInfoListener.toPurchasesDelegate(): PurchasesDelegate =
+    (this as PurchasesDelegateWrapper).wrapped
+
+internal fun PurchasesDelegate.toUpdatedCustomerInfoListener(): UpdatedCustomerInfoListener =
+    PurchasesDelegateWrapper(this)
+
+private class PurchasesDelegateWrapper(val wrapped: PurchasesDelegate) :
+    UpdatedCustomerInfoListener {
+    override fun onReceived(customerInfo: com.revenuecat.purchases.CustomerInfo) {
+        wrapped.onCustomerInfoUpdated(customerInfo.toCustomerInfo())
     }
 }
